Update node1_h1.py
Python script to generate 20 OpenFlow header fields and send them to node h2.
The script will keep trying to establish a connection to node h2 every second until a connection is established or 10 seconds have passed. If a connection is not established within 10 seconds, the script will print an error message and exit.