Update node2_h2.py
Python script to receive the header fields from node h1 and measure the performance on node h2.
The script will keep trying to accept a connection from node h1 every second until a connection is established or 10 seconds have passed. If a connection is not established within 10 seconds, the script will print an error message and exit.